chore: apply extracted empty logic to vectorize
